# docs/swagger.yaml
definitions:
  dto.AccessToken:
    properties:
      access_token:
        type: string
    type: object
  dto.HTTPResp:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  dto.UserCreateReq:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - avatar_url
    - name
    - password
    - phone
    - username
    type: object
  dto.UserDeleteReq:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  dto.UserInfoResp:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      last_logged_in:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
  dto.UserLoginReq:
    properties:
      password:
        type: string
      username_or_email:
        type: string
    required:
    - password
    - username_or_email
    type: object
  dto.UserResp:
    properties:
      avatar_url:
        type: string
      conversation_ids:
        items:
          type: string
        type: array
      created_at:
        type: string
      email:
        type: string
      friends:
        items:
          $ref: '#/definitions/dto.UserInfoResp'
        type: array
      id:
        type: string
      last_logged_in:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.UserUpdateReq:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: Login
      tags:
      - authentications
  /api/auth/logout:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: Logout
      tags:
      - authentications
  /api/user/profile:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: GetProfile
      tags:
      - users
  /api/users:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: DeleteUser
      tags:
      - users
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page of paging
        in: query
        name: page
        type: integer
      - description: size of page of paging
        in: query
        name: page_size
        type: integer
      - description: name of field need to sort
        in: query
        name: sort
        type: string
      - description: sort desc or asc
        in: query
        name: sort_type
        type: string
      - description: keyword to search in model
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: GetUserList
      tags:
      - users
    post:
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: CreateUser
      tags:
      - users
    put:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: UpdateUser
      tags:
      - users
  /api/users/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPResp'
      summary: GetUser
      tags:
      - users
swagger: "2.0"